requirements.txt
# API clients
requests>=2.28.0
google-cloud-speech>=2.17.0
google-cloud-storage>=2.7.0
openai>=1.0.0

# Environment variables
python-dotenv>=0.21.0

# Audio processing
pydub>=0.25.1

# Utilities
tqdm>=4.64.1
